docker_real/requirements.txt
Flask==1.1.2
pytesseract==0.3.6
Pillow==7.2.0
nltk==3.5
pyenchant==3.1.1
opencv-python==4.4.0.44
numpy==1.19.1
imutils==0.5.3
gunicorn==20.0.4
